fix: address PR review findings for bandwidth tracking
- Decouple in/out counter wraparound checks so one direction wrapping
  doesn't silently drop the other direction's valid data
- Prune stale speedCache/lastIface entries for interfaces no longer
  present, preventing unbounded map growth on container hosts
- Cast avg() on bigint columns to float8 in SQL to prevent PostgreSQL
  numeric type being returned as string by the driver
- Fix skipInterface patterns: use strings.HasPrefix, correct isatap/
  Teredo matching
- Compute elapsed time once and share between aggregate and per-
  interface bandwidth calculations

+++ apps/api/src/routes/devices/metrics.ts
@@ -195,8 +195,8 @@ metricsRoutes.get(
         avgDiskUsedGb: sql<number>`avg(${deviceMetrics.diskUsedGb})`,
         totalNetworkIn: sql<bigint>`sum(${deviceMetrics.networkInBytes})`,
         totalNetworkOut: sql<bigint>`sum(${deviceMetrics.networkOutBytes})`,
-        avgBandwidthIn: sql<number>`avg(${deviceMetrics.bandwidthInBps})`,
-        avgBandwidthOut: sql<number>`avg(${deviceMetrics.bandwidthOutBps})`,
+        avgBandwidthIn: sql<number>`avg(${deviceMetrics.bandwidthInBps})::float8`,
+        avgBandwidthOut: sql<number>`avg(${deviceMetrics.bandwidthOutBps})::float8`,
         avgProcessCount: sql<number>`avg(${deviceMetrics.processCount})`
       })
       .from(deviceMetrics)
